dvdnav: fix integer overflow in index

+++ src/input/dvdnav/dvdnav.c
@@ -508,7 +508,8 @@ dvd_open(char *url, char *mode)
 	df->index_size = tmap->nr_of_entries;
 	df->index = calloc(df->index_size, sizeof(*df->index));
 	for(i = 0; i < tmap->nr_of_entries; i++){
-	    df->index[i] = (tmap->map_ent[i] & 0x7fffffff) * DVD_SECTOR_SIZE;
+	    df->index[i] =
+		(uint64_t) (tmap->map_ent[i] & 0x7fffffff) * DVD_SECTOR_SIZE;
 	}
     }
